shell32/tests: Fix out of memory errors and their underlying cause.
The out of memory errors were happening because getstring_test was using
len unitialized when GetString failed. The underlying cause was that
deleting the keys would fail because RegDeleteKey does not work if the
key has subkeys, and one of the keys has a subkey. The test fails when
the keys are still there. The subkey is now deleted before the key that
contains it.

Signed-off-by: Theodore Dubois <tblodt@icloud.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>