wined3d: Prefer shader backends that support both vertex and fragment shaders over ones that support only either.
Apparently there exist some configurations that support both
ARB_vertex_program and ARB_vertex_shader, but only ARB_fragment_program
and not ARB_fragment_shader. Note that this would change the selected
backend for the Geforce 4 cards mentioned in the comment, but hopefully
the mentioned GLSL bug is no longer a concern.

Signed-off-by: Henri Verbeet <hverbeet@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>