wined3d: Fix handling of SM4+ integer registers.
There should be no behavior change for immediate constants and other
registers. After this fix a correct GLSL is generated for SM4+
instructions similar to the following:

mov r0.xyz, vThreadGroupID.xyzx
mov r0.w, vThreadIDInGroupFlattened.x

Previously, it was assumed that an integer register is always in a
position in which an integer type is expected.

Signed-off-by: Józef Kucia <jkucia@codeweavers.com>
Signed-off-by: Henri Verbeet <hverbeet@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>