wined3d: Allow all formats with "glInternal" set in CheckSurfaceCapability().
Rendertarget, depth/stencil and texturing restrictions are applied later in
wined3d_check_device_format(), so to the extent this make a difference it
mostly affects off-screen plain surfaces. Note that if the blitter check were
to be implemented properly it would practically always return TRUE as well,
because of the CPU based blitter.

Signed-off-by: Henri Verbeet <hverbeet@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>