configure: Set PKG_CONFIG_PATH to get the right settings for the target host.
Ideally we would use the right <triplet>-pkg-config tool for the target
host. However Red Hat and SUSE respectively add '-redhat' and '-suse' to
the triplet which makes it hard to find the right tool. Furthermore on
Debian and derivatives a lot of development packages do not support
multiarch so that the required .pc files would not be found.

Setting PKG_CONFIG_PATH side-steps the non-standard Red Hat and SUSE
triplets. It also lets pkg-config fall back to the 'wrong' .pc files
on Debian which is fine in most cases, while using the right one in
cases where it matters (GStreamer 1.0), assuming they are actually
present.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>