d3d11/tests: Run the tests in parallel.
It's hard enough for a single CPU core to keep a modern GPU busy in the
optimal case, but with typical batch sizes of 4 vertices and trivial shaders,
our tests are extremely CPU bound. With 8+-core CPU's being generally
available, there's a lot to be gained. Generally speaking, not all tests can
be run in parallel; tests that touch global state, like e.g. the display mode,
can't. Fortunately, none of the d3d11 tests currently do that.

Unfortunately, this currently doesn't help quite as much on Wine as it does on
Windows. The big wined3d lock prevents a lot of parallelism, even though the
tests use separate Direct3D devices. In the specific case of these tests,
simply disabling the lock is safe, and shows how much could be gained by
simply pushing the lock down a layer to the device.

Signed-off-by: Henri Verbeet <hverbeet@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>